requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking
#
attrs==22.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.30
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.38
    # via aws-cdk-lib
aws-cdk-aws-lambda-python-alpha==2.54.0a0
    # via -r requirements.in
aws-cdk-lib==2.54.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
boto3==1.26.142
    # via -r requirements.in
botocore==1.29.142
    # via
    #   boto3
    #   s3transfer
cattrs==22.2.0
    # via jsii
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
constructs==10.1.189
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
exceptiongroup==1.0.4
    # via cattrs
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.72.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
pycryptodome==3.18.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
requests==2.31.0
    # via -r requirements.in
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.4.0
    # via jsii
urllib3==1.26.16
    # via
    #   botocore
    #   requests
